+++ proj_start.bash
@@ -1,1 +1,2 @@
-alias start='uvicorn app.main:app --reload'
+alias start_backend='uvicorn app.main:app --reload'
+alias start_frontend='cd frontend && npm run dev'